Found the problematic factor of 2: it was in my math reasoning, rather than in the implementation. Also fixed another mistake.
The issue was with the "symmetrized" interaction tensors: I shouldn't have been halving them in my code, since I was carrying a factor of 1/2 in my notes.

I had also forgotten a factor of 1/2 in the classical energy, so I was accidentally getting the right classical energy despite the above mistake.