More tests carried out on honeycomb FM with DM interaction. Only somewhat conclusive because of apparent errors in the papers (see below).
According to my own calculations, the Hamiltonian from arXiv:2106.01430 appears to be missing a factor of 2 in front of the DM interaction. Once this
mistake is fixed, the analytical expressions agree with the numerics.

According to my own calculations, the Hamiltonian from J. Phys.: Condens. Matter 28 386001 (2016) appears to have two mistakes in its expression for
h_z after Eq. 12: the term 6nu_s' is missing, and there is a spurious prefactor of 2 in front of nu_t. With these mistakes fixed, the numerics agree
with the analytical expressions.